fix: handle github provider not supporting listing
instead reporting an error and throwing one, made it so that listing only throws an error, which can be handled anyhow later on. In cleanup, will handle it by writing an info log and exiting from clean up early. This was done to mitigate the errors reported from github, every time the cleanup would run and the code would reach the changed lines, which happened (in my case) when "max-age" was specified for the inputs.